implement recursive graph* runtime with routing
includes condition checks and action handling (subroutines included)

It wasn't really a PDA (or DFA) at any time due to the pathfinding
requirement (in state u, need to go to state v) that these FSA don't
have. Though the subroutine with stack is sort of inspired by PDAs.

The subroutines are more similar to a kind of recursive graph with
many duplicate subgraphs.